+++ modules/apps/message-boards/message-boards-comment/src/com/liferay/message/boards/comment/MBCommentImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.comment.service;
+package com.liferay.message.boards.comment;
 
 import com.liferay.portal.kernel.comment.WorkflowableComment;
 import com.liferay.portal.model.User;
