+++ modules/apps/message-boards/message-boards-comment/src/com/liferay/message/boards/comment/MBDiscussionPermissionImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.comment.service;
+package com.liferay.message.boards.comment;
 
 import com.liferay.portal.kernel.comment.BaseDiscussionPermission;
 import com.liferay.portal.kernel.exception.PortalException;
