+++ modules/apps/message-boards/message-boards-comment/src/com/liferay/message/boards/comment/MBDiscussionStagingHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.comment.service;
+package com.liferay.message.boards.comment;
 
 import com.liferay.portal.kernel.comment.DiscussionStagingHandler;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
